python/requirements.txt
bottle==0.12.8
Pillow==2.8.1
